128MB of more memory for the CPU
We've discussed this: even 256MB of GPU memory is overkill - this was all caused by the fact that this had been set to insanely high levels when we started to review the config.txt settings. According to system telemetry data during runtime, it is very hard to even reach or exceed 100MB of GPU mem usage. Tested this on 1080p solution with YouTube SD videos running, a flash SD video running at the same time, while having Cog open in the foreground. GPU mem usage was consistently lower than even 90MB. Having these 128MB of additional RAM available for the actual system instead of keeping it reserved for the GPU which will never use it is the sane decision here.